Only download cdh tarball if we do not already have it.

+++ build.sh
@@ -118,8 +118,10 @@ build_maven_project "camus" "${GIT_MODE}confluentinc/camus.git" "package" "confl
 if [ ! -d hadoop-cdh ]; then
     hadoop_cdh_version="hadoop-2.5.0-cdh5.3.0"
     if [ ! -e $hadoop_cdh_version ]; then
-        curl http://archive.cloudera.com/cdh5/cdh/5/$hadoop_cdh_version.tar.gz -o "$hadoop_cdh_version.tar.gz"
+        if [ ! -e ${hadoop_cdh_version}.tar.gz ]; then
+            curl http://archive.cloudera.com/cdh5/cdh/5/$hadoop_cdh_version.tar.gz -o "$hadoop_cdh_version.tar.gz"
+        fi
+        tar xvzf ${hadoop_cdh_version}.tar.gz
     fi
-    tar xvzf $hadoop_cdh_version.tar.gz
     mv $hadoop_cdh_version hadoop-cdh
 fi
